+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Improvements
 * Updated InspectorMessage reporting for PyNWB read errors to improve readability [#603](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/603)
+* Added support for PyNWB 3.1 and NWB Schema 2.9 [#602](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/602)
 
 # v0.6.3 (March 13, 2025)